+++ app/src/main/assets/fhconfig.properties
@@ -1,6 +1,6 @@
 host = https://testing.feedhenry.me
-appid = l5z254qjdsbx7zpk3fuibsfr
+appid = l5z254um56q52pbqnmv5sqzf
 projectid = l5z254vsghfpliwin3z5v4zo
-appkey = 84355d38a6182c789210ee2bf17c7bf3c3498807
-connectiontag = 0.0.4
+appkey = 9e94fbb05e71f201754a6d1de205cb3524b6b37
+connectiontag = 0.0.8
