+++ app/src/main/res/values/colors.xml
@@ -7,9 +7,8 @@
     If you do not have a contract, you do not have a license to use this software.
 
 -->
-<TextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
-    android:padding="10dp"
-    android:textSize="20sp" >
-</TextView>
+<resources>
+    <color name="redhat_red">#CC0000</color>
+    <color name="redhat_dark_red">#781f1c</color>
+    <color name="redhat_accent">#564979</color>
+</resources>
