+++ app/src/main/res/values/strings.xml
@@ -12,5 +12,5 @@
     <string name="app_name">SyncSample</string>
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
-
+    <string name="sync_demo_title">Sync Demo</string>
 </resources>
