+++ app/src/main/res/values/styles.xml
@@ -10,7 +10,30 @@
 
   <!-- Base application theme. -->
   <style name="AppTheme" parent="Theme.AppCompat.Light.DarkActionBar">
-    <!-- Customize your theme here. -->
+
+      <item name="colorPrimary">@color/redhat_red</item>
+      <item name="colorPrimaryDark">@color/redhat_dark_red</item>
+      <item name="colorAccent">@color/redhat_accent</item>
   </style>
+    <style name="AppTheme.NoActionBar">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+    </style>
+    <style name="AppTheme.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" >
+        <item name="android:fontFamily">sans-serif-condensed-light</item>
+        <item name="android:textSize">8sp</item>
+    </style>
+    <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" >
+    </style>
+
+    <style name="ToolbarTheme.Title" parent="@style/TextAppearance.Widget.AppCompat.Toolbar.Title">
+        <item name="android:fontFamily">sans-serif-condensed-light</item>
+        <item name="android:textSize">16sp</item>
+    </style>
+
+    <style name="FloatingActionButton">
+
+    </style>
+
 
 </resources>
